change setup
